// src/interfaces/StatsAward.ts
interface StatsAward {
    week: number;
    sixmonth: number;
    year: number;
    january: number;
    february: number;
    march: number;
    april: number;
    may: number;
    june: number;
    july: number;
    august: number;
    september: number;
    october: number;
    november: number;
    december: number;
}

export default StatsAward;